+++ dio/lib/src/compute/compute.dart
@@ -25,7 +25,9 @@
 
 import 'dart:async';
 
-import 'compute_io.dart' if (dart.library.html) 'compute_web.dart' as _c;
+import 'compute_io.dart'
+    if (dart.library.js_interop) 'compute_web.dart'
+    if (dart.library.html) 'compute_web.dart' as _c;
 
 /// Signature for the callback passed to [compute].
 ///
