+++ dio/lib/src/multipart_file/browser_multipart_file.dart
@@ -1,21 +1,2 @@
-import 'package:http_parser/http_parser.dart';
-
-final _err = UnsupportedError(
-  'MultipartFile is only supported where dart:io is available.',
-);
-
-Never multipartFileFromPath(
-  String filePath, {
-  String? filename,
-  MediaType? contentType,
-  final Map<String, List<String>>? headers,
-}) =>
-    throw _err;
-
-Never multipartFileFromPathSync(
-  String filePath, {
-  String? filename,
-  MediaType? contentType,
-  final Map<String, List<String>>? headers,
-}) =>
-    throw _err;
+export 'package:dio_web_adapter/dio_web_adapter.dart'
+    show multipartFileFromPath, multipartFileFromPathSync;
