+++ dio/pubspec.yaml
@@ -25,6 +25,8 @@ dependencies:
   meta: ^1.5.0
   path: ^1.8.0
 
+  dio_web_adapter: '>=1.0.0 <3.0.0'
+
 dev_dependencies:
   lints: any
   test: ^1.5.1
