+++ dio_workspace.iml
@@ -22,6 +22,9 @@
       <excludeFolder url="file://$MODULE_DIR$/plugins/native_dio_adapter/example/.dart_tool" />
       <excludeFolder url="file://$MODULE_DIR$/plugins/native_dio_adapter/example/.pub" />
       <excludeFolder url="file://$MODULE_DIR$/plugins/native_dio_adapter/example/build" />
+      <excludeFolder url="file://$MODULE_DIR$/plugins/web_adapter/.dart_tool" />
+      <excludeFolder url="file://$MODULE_DIR$/plugins/web_adapter/.pub" />
+      <excludeFolder url="file://$MODULE_DIR$/plugins/web_adapter/build" />
     </content>
     <orderEntry type="sourceFolder" forTests="false" />
     <orderEntry type="library" name="Dart SDK" level="project" />
