+++ plugins/web_adapter/analysis_options.yaml
@@ -1,0 +1,5 @@
+include: ../../analysis_options.yaml
+
+linter:
+  rules:
+    implementation_imports: false
