+++ plugins/web_adapter/test/web_adapter_test.dart
@@ -1,0 +1,6 @@
+// import 'package:dio_web_adapter/dio_web_adapter.dart';
+import 'package:test/test.dart';
+
+void main() {
+  group('A group of tests', () {});
+}
